EDKIIUS-25: Fix the VS build error
Fix the VS build error

+++ DynamicTablesPkg/Library/Common/AmlLib/CodeGen/AmlResourceDataCodeGen.c
@@ -870,7 +870,7 @@ AmlCodeGenRdWordBusNumber (
            (UINT16) AddressMinimum,
            (UINT16) AddressMaximum,
            (UINT16) AddressTranslation,
-           RangeLength,
+           (UINT16) RangeLength,
            ResourceSourceIndex,
            ResourceSource,
            NameOpNode,
